requirements.txt
absl-py==1.3.0
action-tutorials-py==0.9.3
ament-copyright==0.9.6
ament-cppcheck==0.9.6
ament-cpplint==0.9.6
ament-flake8==0.9.6
ament-index-python==1.1.0
ament-lint==0.9.6
ament-lint-cmake==0.9.6
ament-package==0.9.5
ament-pep257==0.9.6
ament-uncrustify==0.9.6
ament-xmllint==0.9.6
astunparse==1.6.3
bidict==0.21.2
Brotli==1.0.9
cachetools==5.2.0
captum==0.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.1
cloudpickle==2.2.0
coloredlogs==15.0.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
deeplift==0.6.13.0
demo-nodes-py==0.9.3
dm-sonnet==2.0.0
dm-tree==0.1.7
dnspython==2.2.1
domain-coordinator==0.9.2
eventlet==0.33.0
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
Flask==2.0.1
Flask-Compress==1.10.1
Flask-SocketIO==5.1.1
flatbuffers==2.0.7
fonttools==4.38.0
gast==0.4.0
gdbgui==0.15.1.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.50.0
h5py==3.7.0
humanfriendly==10.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.6.8
lark==1.1.4
launch==0.10.8
launch-ros==0.11.6
launch-testing==0.10.8
launch-testing-ros==0.11.6
launch-xml==0.10.8
launch-yaml==0.10.8
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.6.2
mpmath==1.2.1
networkx==2.8.8
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.12.0
onnx-tf==1.10.0
onnx2keras==0.0.24
onnxruntime==1.13.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
osrf-pycommon==0.1.11
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.19.6
psutil==5.9.4
pt2keras==0.0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygdbmi==0.10.0.1
Pygments==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
pytorch2keras==0.2.4
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
quality-of-service-demo-py==0.9.3
requests==2.28.1
requests-oauthlib==1.3.1
ros2action==0.9.11
ros2bag==0.3.9
ros2cli==0.9.11
ros2component==0.9.11
ros2doctor==0.9.11
ros2interface==0.9.11
ros2launch==0.11.6
ros2lifecycle==0.9.11
ros2multicast==0.9.11
ros2node==0.9.11
ros2param==0.9.11
ros2pkg==0.9.11
ros2run==0.9.11
ros2service==0.9.11
ros2topic==0.9.11
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt==1.1.2
rqt-action==0.4.9
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-moveit==1.0.1
rqt-msg==1.0.5
rqt-plot==1.1.1
rqt-publisher==1.1.3
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.5
rqt-robot-steering==1.0.0
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.2
rqt-top==1.0.2
rqt-topic==1.2.2
rsa==4.9
saliency==0.2.0
scikit-image==0.18.3
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sros2==0.9.5
sympy==1.11.1
tabulate==0.9.0
teleop-twist-keyboard==2.3.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.18.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.18.0
termcolor==2.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
topic-monitor==0.9.3
torch==1.13.0
torchvision==0.14.0
typeguard==2.13.3
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.0.0
wrapt==1.14.1
zipp==3.10.0
